+++ Sprint_0/Interfaces/IEnemy.cs
@@ -8,8 +8,8 @@ public interface IEnemy
 {
     //required type properties
     public IEnemyState state { get; set; }
-    public int xPos { get; set; }
-    public int yPos { get; set; }
+    public float xPos { get; set; }
+    public float yPos { get; set; }
     public int health { get; set; }
     public int randTime { get; set; }
 
@@ -27,6 +27,6 @@ public interface IEnemy
 
     int getHeight();
     int getWidth();
-    int getSpeed();
+    float getSpeed();
 
 }
