+++ Sprint_0/Items/ItemClasses/Bow.cs
@@ -9,8 +9,8 @@ using System.Reflection.Metadata;
 public class Bow : IItem
 {
     //item properties
-    private int xPos;
-    private int yPos;
+    private float xPos;
+    private float yPos;
     private int Width = 24;
     private int Height = 48;
     private ItemManager man;
@@ -32,7 +32,7 @@ public class Bow : IItem
     {
         Texture2D bow = Texture2DStorage.GetItemSpritesheet();
         Rectangle sourceRect = ItemRectStorage.getBowSprite();
-        Rectangle destRect = new Rectangle(this.xPos, this.yPos, this.Width, this.Height);
+        Rectangle destRect = new Rectangle((int)this.xPos, (int)this.yPos, this.Width, this.Height);
         spriteBatch.Draw(bow, destRect, sourceRect, Color.White);
 
     }
@@ -48,12 +48,12 @@ public class Bow : IItem
         return this.Width;
     }
 
-    public int getX()
+    public float getX()
     {
         return this.xPos;
     }
 
-    public int getY()
+    public float getY()
     {
         return this.yPos;
     }
