+++ Sprint_0/Items/ItemClasses/DaCoin.cs
@@ -9,8 +9,8 @@ using System.Reflection.Metadata;
 public class DaCoin : IItem
 {
     //item properties
-    private int xPos;
-    private int yPos;
+    private float xPos;
+    private float yPos;
     private int Width = 48;
     private int Height = 48;
 
@@ -20,7 +20,7 @@ public class DaCoin : IItem
 
     private ItemManager man;
 
-    public DaCoin(ItemManager manager, int xPosition, int yPosition)
+    public DaCoin(ItemManager manager, float xPosition, float yPosition)
     {
         this.xPos = xPosition + 8;
         this.yPos = yPosition + 8;
@@ -40,7 +40,7 @@ public class DaCoin : IItem
     {
         Texture2D rupee = Texture2DStorage.GetItemSpritesheet();
         Rectangle sourceRect = ItemRectStorage.getDaCoinSourceRect();
-        Rectangle destRect = new Rectangle(this.xPos, this.yPos, this.Width, this.Height);
+        Rectangle destRect = new Rectangle((int)this.xPos, (int)this.yPos, this.Width, this.Height);
         spriteBatch.Draw(rupee, destRect, sourceRect, Color.White);
 
     }
@@ -56,12 +56,12 @@ public class DaCoin : IItem
         return this.Width;
     }
 
-    public int getX()
+    public float getX()
     {
         return this.xPos;
     }
 
-    public int getY()
+    public float getY()
     {
         return this.yPos;
     }
