+++ Sprint_0/Items/ItemClasses/HeartContainer.cs
@@ -9,8 +9,8 @@ using System.Reflection.Metadata;
 public class HeartContainer : IItem
 {
     //item properties
-    private int xPos;
-    private int yPos;
+    private float xPos;
+    private float yPos;
     private int Width = 24;
     private int Height = 48;
 
@@ -34,7 +34,7 @@ public class HeartContainer : IItem
     {
         Texture2D heartContainer = Texture2DStorage.GetItemSpritesheet();
         Rectangle sourceRect = ItemRectStorage.getHeartContainerSprite();
-        Rectangle destRect = new Rectangle(this.xPos, this.yPos, this.Width, this.Height);
+        Rectangle destRect = new Rectangle((int)this.xPos, (int)this.yPos, this.Width, this.Height);
         spriteBatch.Draw(heartContainer, destRect, sourceRect, Color.White);
 
     }
@@ -50,12 +50,12 @@ public class HeartContainer : IItem
         return this.Width;
     }
 
-    public int getX()
+    public float getX()
     {
         return this.xPos;
     }
 
-    public int getY()
+    public float getY()
     {
         return this.yPos;
     }
