+++ Sprint_0/NPCs/Enemies/Enemy Classes/AdamSandler.cs
@@ -14,8 +14,8 @@ public class AdamSandler : IEnemy
 
     /* Properties that change, the heart of the enemy*/
     public IEnemyState state {  get;  set; }
-    public int xPos { get; set; }
-    public int yPos { get; set; }
+    public float xPos { get; set; }
+    public float yPos { get; set; }
     public int health { get; set; }
     public int randTime { get; set; }
 
@@ -157,7 +157,7 @@ public class AdamSandler : IEnemy
         return width;
     }
 
-    public int getSpeed()
+    public float getSpeed()
     {
         return enemySpeed;
     }
