+++ Sprint_0/NPCs/Enemies/Enemy Sprites/AdamSandlerGolfBallSprite.cs
@@ -56,9 +56,9 @@ public class AdamSandlerGolfBallSprite : IEnemySprite
         //Does not take damage.
     }
 
-    public void update(int xPos, int yPos, int facingDirections, int time)
+    public void update(float xPos, float yPos, int facingDirections, int time)
     {
-        this.destinationRectangle = new Rectangle(xPos, yPos, 16, 16);
+        this.destinationRectangle = new Rectangle((int)xPos, (int)yPos, 16, 16);
         
         frame = time % 4;
     }
