+++ Sprint_0/NPCs/Enemies/Enemy Sprites/GoriyaBoomerangSprite.cs
@@ -50,9 +50,9 @@ public class GoriyaBoomerangSprite : IEnemySprite
         //Does not take damage.
     }
 
-    public void update(int xPos, int yPos, int facingDirections, int time)
+    public void update(float xPos, float yPos, int facingDirections, int time)
     {
-        this.destinationRectangle = new Rectangle(xPos, yPos, 32, 32);
+        this.destinationRectangle = new Rectangle((int)xPos, (int)yPos, 32, 32);
         
         frame = time % 3;
     }
