+++ Sprint_0/NPCs/Enemy Projectiles & more/AdamSandlerGolfBall.cs
@@ -9,8 +9,8 @@ public class AdamSandlerGolfBall : IEnemy
 {
     
     public IEnemyState state { get; set; }
-    public int xPos { get; set; }
-    public int yPos { get; set; }
+    public float xPos { get; set; }
+    public float yPos { get; set; }
     public int health { get; set; }
     public int randTime { get; set; }
 
@@ -19,7 +19,7 @@ public class AdamSandlerGolfBall : IEnemy
     //8x8 on spritesheet
     private const int height = 16;
     private const int width = 16;
-    private const int enemySpeed = 30;
+    private const float enemySpeed = 30;
     private int timeUntilDeath;
     private EnemyManager man;
     private int golfBall;
@@ -129,7 +129,7 @@ public class AdamSandlerGolfBall : IEnemy
         return width;
     }
 
-    public int getSpeed()
+    public float getSpeed()
     {
         return enemySpeed;
     }
