+++ Sprint_0/NPCs/Enemy Projectiles & more/DeathAnimation.cs
@@ -9,8 +9,8 @@ public class DeathAnimation : IEnemy
 {
 
     public IEnemyState state { get; set; }
-    public int xPos { get; set; }
-    public int yPos { get; set; }
+    public float xPos { get; set; }
+    public float yPos { get; set; }
     public int health { get; set; }
     public int randTime { get; set; }
 
@@ -19,7 +19,7 @@ public class DeathAnimation : IEnemy
 
     private const int height = 16;
     private const int width = 16;
-    private const int enemySpeed = 0;
+    private const float enemySpeed = 0;
     private EnemyManager man;
 
 
@@ -95,7 +95,7 @@ public class DeathAnimation : IEnemy
         return width;
     }
 
-    public int getSpeed()
+    public float getSpeed()
     {
         return enemySpeed;
     }
